+++ CHANGES.rst
@@ -5,6 +5,7 @@ IP Navigator CHANGES
 development
 ===========
 - Upgrade to MechanicalSoup 0.10.0
+- Improve whitelabel capabilities
 
 2018-01-22 0.148.1
 ==================